pyproject.toml
[tool.poetry]
name = "pyproject.toml"
version = "1"
description = "job portal backend"
authors = ["vishnu"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.30.1"
fastapi = "0.111.0"
redis = "^5.0.7"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
aiohttp = "^3.9.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
beanie = "^1.26.0"
aiofiles = "^24.1.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
